requirements.txt
typer>=0.12.0
rich>=13.7.0
ruamel.yaml>=0.18.6
gitpython>=3.1.43
watchdog>=4.0.1
pytest>=8.0.0
